Omit unnecessary unboxing calls in decompiled code

+++ plugins/java-decompiler/engine/testData/obfuscated/aj.java
@@ -105,7 +105,7 @@ public class aj implements am {
                Entry var6 = (Entry)var5.next();
                Long var7 = (Long)this.b.get(var6.getKey());
                if (var7 != null) {
-                  double var8 = (double)((((Long)var6.getValue()).longValue() - var7.longValue()) * 10L) / (double)var3;
+                  double var8 = (double)(((Long)var6.getValue() - var7) * 10L) / (double)var3;
                   var1.a((Object)(new ar(h[1], (String)var6.getKey(), "%", var8 * 100.0D)));
                }
             }
